+++ aawedha/io/san_diego.py
@@ -151,8 +151,8 @@ class SanDiego(DataSet):
             X.append(eeg.reshape((samples, channels, blocks * targets), order='F'))
             Y.append(y)
 
-        X = np.array(X)
-        Y = np.array(Y).squeeze()
+        X = np.array(X, dtype=np.float32)
+        Y = np.array(Y, dtype=np.float32).squeeze()
         return X, Y
 
     def download_raw(self, store_path=None):
