+++ aawedha/io/tsinghua.py
@@ -173,8 +173,8 @@ class Tsinghua(DataSet):
             del eeg
             del y
 
-        X = np.array(X)
-        Y = np.array(Y).squeeze()
+        X = np.array(X, dtype=np.float32)
+        Y = np.array(Y, dtype=np.float32).squeeze()
         return X, Y
 
     def download_raw(self, store_path=None):
